Move team creation to repository
